[#4] delete mypage recycler view
