--- app/src/main/java/com/example/knucseapp/ui/mypage/Menuname.kt
@@ -1,8 +1,0 @@
-package com.example.knucseapp.ui.mypage
-
-class Menuname {
-    companion object {
-        val grouplist = listOf("예약 내역", "제재 내역", "내가 쓴 글", "내가 쓴 댓글", "환경설정", "로그아웃")
-        val setting = listOf("푸시 알람", "소리")
-    }
-}
